refactor : 파일 및 gradle 정리
