+++ src/main/java/project/backend/user/WebConfig.java
@@ -1,4 +1,4 @@
-package project.backend.config;
+package project.backend.user;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
